Split the Storage tests into different modules
FS and ImmutableDB are now separate modules. A Util module was also
introduced.
