+++ ouroboros-consensus/ouroboros-consensus.cabal
@@ -246,8 +246,11 @@ test-suite test-storage
   default-language: Haskell2010
   main-is:          Main.hs
   other-modules:
-                    Test.Ouroboros.StateMachine
                     Test.Ouroboros.Storage
+                    Test.Ouroboros.Storage.Util
+                    Test.Ouroboros.Storage.FS
+                    Test.Ouroboros.Storage.FS.StateMachine
+                    Test.Ouroboros.Storage.ImmutableDB
                     Test.Ouroboros.Storage.ImmutableDB.Sim
                     Test.Util.RefEnv
   build-depends:    base,
